backend/rag/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
chromadb
sentence-transformers
transformers
flask
flask_cors
torch
pyMySql[rsa]
peewee
python-dotenv
redis
PyJWT
waitress
requests
google-genai
cryptography
python-dateutil
beautifulsoup4
